requirements.txt
streamlit==1.31.0
pymongo==4.6.1
requests==2.31.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
